Replace wrappers with a simple extension module
