--- wearableinterface/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/wearableinterface/CapabilityManager.kt
@@ -1,9 +1,0 @@
-package com.boswelja.smartwatchextensions.wearableinterface
-
-interface CapabilityManager {
-    fun addLocalCapability(capability: String)
-
-    fun removeLocalCapability(capability: String)
-
-    suspend fun getCapabilitiesFor(watchId: String): Set<String>
-}
