--- wearableinterface/wear/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest />
