--- wearableinterface/wear/src/main/kotlin/com/boswelja/smartwatchextensions/wearableinterface/CapabilityManager.kt
@@ -1,7 +1,0 @@
-package com.boswelja.smartwatchextensions.wearableinterface
-
-interface CapabilityManager {
-    fun addLocalCapability(capability: String)
-
-    fun removeLocalCapability(capability: String)
-}
